shuffle around code a bit, per PR feedback
